Rewrite TowardsGoalBy as StepTowards:  * A proper name, as what this function does is exactly that:   - It moves a given value with a given step towards a given goal  * Change the code to be more clear about what it means:   - e.g. instead of (step < goal - val) use (step + val < goal) to determine whether (step + val) is a valid return value.  * Make StepTowards a template function to allow the compiler to perform better optimisations   - E.g. better optimisations than when using doubles for values that are stored in floats
git-svn-id: https://svn.il.fontys.nl/usersvn/fishtank/svn/trunk@386 bb1b31de-baa9-489e-af1c-03370f2903b8

+++ fishtank/Vis.h
@@ -24,7 +24,7 @@ class Vis : public Object
 		double speed;
 		double turn_speed;
 		double bending;
-		double pitch;
+		float pitch;
 
 		//instellingen
 		double max_speed;
